add tox.ini and lint rule, fix pep8 errors
This adds a tox.ini base file for adding testing to pyunifi, and
starts with a simple pep8 check. This uncovered at least 2 programming
errors in controller.py because of symbols that did not exist.

This can be run with ``tox -e lint``

Note: urlencode exists in different places in python2 and python3, so
this makes this python3 only right now. As home assistant is the
primary consumer, that seems fine.